# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Todo API
host: localhost:10010
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /:
    get:
      description: "A description: This endpoint returns all todos in database"
      summary: "A shorter summary of the operation"
      operationId: "GetAllTodos"
      parameters: []
      responses:
        "200":
          description: "an array of todos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Todo"
      x-swagger-router-controller: "GetAllTodos"
    post:
      description: 'Add a new Todo'
      operationId: 'AddTodo'
      produces:
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'todo'
          description: 'the Todo to be added'
          required: true
          schema:
            $ref: "#/definitions/Todo"
      responses:
        '201':
          description: 'Successful Todo add'
          schema:
            $ref: "#/definitions/Todo"
      x-swagger-router-controller: "AddTodo"
...
definitions:
  Todo:
    type: object
    properties:
      todo_id:
        type: "integer"
        description: "Id for the todo"
      todo:
        type: "string"
        description: "The todo itself"
      datecreated:
        type: "string"
        format: "date-time"
        description: "The date the todo was created. Set by server"
      author:
        type: "string"
        description: "Name of the person who created the todo"
      duedate:
        type: "string"
        format: "date-time"
        description: "Date the todo is due"
      completed:
        type: "boolean"
        description: "Indicate if the todo is completed or not"